feat: update tournament card for all pages
