+++ src/app/modules/tournament-management/pages/administrator-tournaments/administrator-tournaments.component.html
@@ -4,31 +4,9 @@
 		(filtersChanged)="onFiltersChanged($event)">
 	</app-tournament-filters>
 
-	<div class="all-tournaments" *ngIf="allTournaments.length > 0">
-		<div class="tournament" *ngFor="let tournament of filteredTournaments">
-			<div class="left">
-				<div class="column">
-					<div class="id" matTooltip="Tournament import id">
-						{{ tournament.id }}
-					</div>
-
-					<div>
-						{{ tournament.name }}
-					</div>
-				</div>
-
-				<div class="column">
-					Submitted by {{ tournament.createdBy.username }}
-				</div>
-			</div>
-
-			<div class="right">
-				<div class="button">
-					<button mat-mini-fab color="accent" matTooltip="Edit this tournament" [matTooltipPosition]="'above'" (click)="editTournament(tournament)"><mat-icon>edit</mat-icon></button>
-					<button mat-mini-fab color="primary" matTooltip="Import this tournament" [matTooltipPosition]="'above'" (click)="importTournament(tournament)"><mat-icon>cloud_download</mat-icon></button>
-					<button mat-mini-fab color="warn" matTooltip="Delete this tournament" [matTooltipPosition]="'above'" (click)="deleteTournament(tournament)"><mat-icon>delete</mat-icon></button>
-				</div>
-			</div>
-		</div>
+	<div class="all-tournaments">
+		<ng-container *ngIf="allTournaments.length > 0">
+			<app-tournament-card [tournament]="tournament" cardType="administrator" *ngFor="let tournament of filteredTournaments"></app-tournament-card>
+		</ng-container>
 	</div>
 </div>
