+++ src/app/modules/tournament-management/pages/import-tournament/import-tournament.component.scss
@@ -4,66 +4,8 @@
 	padding: 8px;
 
 	.all-tournaments {
-		.tournament {
-			display: flex;
-			flex: 0 1 auto;
-			justify-content: space-between;
-
-			border: 1px solid darken(#3c4452, 15%);
-
-			-webkit-box-shadow: 0px 0px 15px 0px rgba(0, 0, 0, 0.75);
-			-moz-box-shadow: 0px 0px 15px 0px rgba(0, 0, 0, 0.75);
-			box-shadow: 0px 0px 15px 0px rgba(0, 0, 0, 0.75);
-
-			margin-bottom: 8px;
-			padding: 16px;
-
-			.left {
-				display: flex;
-				flex: 1;
-				flex-direction: column;
-				justify-content: center;
-
-				.column {
-					display: flex;
-
-					.id {
-						background-color: $primary;
-						color: black;
-
-						border-radius: 9px;
-
-						padding: 2px 16px 2px 16px;
-						margin-right: 8px;
-
-						cursor: default;
-					}
-
-					* {
-						display: flex;
-						align-self: center;
-					}
-				}
-			}
-
-			.right {
-				display: flex;
-				flex: 1;
-				justify-content: flex-end;
-
-				.button {
-					display: flex;
-					align-items: center;
-				}
-			}
-
-			&:hover {
-				border: 1px solid white;
-
-				-webkit-box-shadow: 0px 0px 15px 0px rgba(255, 255, 255, 0.75);
-				-moz-box-shadow: 0px 0px 15px 0px rgba(255, 255, 255, 0.75);
-				box-shadow: 0px 0px 15px 0px rgba(255, 255, 255, 0.75);
-			}
-		}
+		display: flex;
+		flex-direction: column;
+		gap: 8px;
 	}
 }
